requirements.txt
Django==1.9
Fabric==1.10.0
Jinja2==2.9.6
Markdown==2.6.9
MarkupSafe==1.0
PyNaCl==1.1.2
Pygments==2.2.0
Sphinx==1.2.3
amqp==2.2.1
apiai==1.2.3
argparse==1.2.1
asn1crypto==0.22.0
bcrypt==3.1.3
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
cryptography==2.0.3
django-appconf==1.0.2
django-celery==3.1.16
django-cleanup==0.1.12
django-compressor==1.4
django-configurations==0.8
django-environ==0.4.3
django-filter==1.0.4
django-filters==0.2.1
djangorestframework==3.6.3
docutils==0.14
enum34==1.1.6
google-api-python-client==1.6.2
gspread==0.6.2
httplib2==0.10.3
idna==2.6
ipaddress==1.0.18
kombu==4.1.0
numpy==1.13.1
oauth2client==4.1.2
paramiko==2.2.1
psycopg2==2.5.4
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
python-json-logger==0.1.2
pytz==2017.2
redis==2.10.6
requests==2.18.4
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
wsgiref==0.1.2
